pyproject.toml
[project]
name = "graphrag-hybrid"
version = "0.1.0"
description = "Neo4j と Qdrant を活用したハイブリッド GraphRAG システム"
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    # Neo4j dependencies
    "neo4j>=5.11.0",
    # Qdrant dependencies
    "qdrant-client>=1.5.0",
    # NLP and embedding dependencies
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    # MCP Server
    "fastmcp>=2.0.0",
    # Utilities
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "uuid>=1.30",
    # UI & Document Processing
    "streamlit>=1.30.0",
    "pymupdf4llm>=0.0.10",
    # LLM Entity Extraction
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
